# api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "LoanLens API",
    "description": "Advanced Loan Default Prediction API with explainable AI, model management, A/B testing, drift detection, and automated retraining.",
    "version": "2.0.0",
    "contact": {
      "name": "LoanLens Development Team",
      "email": "support@loanlens.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.loanlens.ai/v2",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/predict": {
      "post": {
        "summary": "Make Prediction",
        "description": "Makes a single loan default prediction with probability and SHAP explanations",
        "operationId": "predict",
        "tags": ["Predictions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing features or no models loaded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PredictionRequest": {
        "type": "object",
        "required": ["features"],
        "properties": {
          "features": {
            "type": "object",
            "description": "Feature values for prediction"
          },
          "model_name": {"type": "string"},
          "threshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "default": 0.5
          }
        }
      },
      "PredictionResponse": {
        "type": "object",
        "properties": {
          "prediction": {"type": "integer"},
          "probability": {"type": "number"},
          "model_name": {"type": "string"},
          "threshold": {"type": "number"},
          "timestamp": {"type": "string"},
          "explanation": {"type": "object"},
          "latency_ms": {"type": "number"}
        }
      },
...
    }
  }
}